dockerized the rails project and updated ruby version to 3.0.7

+++ db/migrate/20250410093208_drop_users_table.rb
@@ -1,5 +1,16 @@
 class DropUsersTable < ActiveRecord::Migration[6.1]
-  def change
-    drop_table :users
+  def up
+    drop_table :users if table_exists?(:users)
+  end
+
+  def down
+    create_table :users do |t|
+      # Define the columns you would need if you were to recreate this table
+      # For example:
+      # t.string :name
+      # t.string :email
+      
+      t.timestamps
+    end
   end
 end
